requirements.txt
allure-pytest==2.8.11
allure-python-commons==2.8.11
attrs==19.3.0
configparser==4.0.2
pytest==5.4.1
pytest-assume==2.2.0
selene==1.0.2
selenium==3.141.0
urllib3==1.25.8
